refactor: move logging config to runtime (#863)
